Fixed signed and unsigned comparison in AFK timer

+++ brick/brick_app_linux.cc
@@ -26,8 +26,8 @@
 namespace {
   const int kMainWindowWidth = 914;
   const int kMainWindowHeight = 454;
-  const int32 kIdleTimeout = 600000L;
-  const unsigned long kIdleCheckInterval = 10000;
+  const unsigned long kIdleTimeout = 600000L;
+  const unsigned long kIdleCheckInterval = 10000L;
   std::string kAppIcons[] = {"brick16.png", "brick32.png", "brick48.png", "brick128.png", "brick256.png"};
   std::string szWorkingDir;  // The current working directory
   std::string startupLocale;
